Rename files s/dictionariesSchema/thesaurusSchema
