+++ app/shared/react-dnd.d.ts
@@ -1,3 +1,4 @@
 /** @format */
 
+// TODO: Update react-dnd to a newer version that doesn't require an explicit type declaration.
 declare module 'react-dnd';
