+++ app/shared/thesaurusType.d.ts
