+++ database/sync-topic-classification.ts
@@ -1,6 +1,5 @@
 /** @format */
 /* eslint-disable no-await-in-loop, no-console, camelcase */
-
 import { tcServer, useThesaurusNames } from 'api/config/topicclassification';
 import entities from 'api/entities';
 import { QueryForEach, WithId } from 'api/odm';
@@ -13,10 +12,11 @@ import JSONRequest from 'shared/JSONRequest';
 import { propertyTypes } from 'shared/propertyTypes';
 import { ensure, sleep } from 'shared/tsUtils';
 import yargs from 'yargs';
+
 import { MetadataObject } from '../app/api/entities/entitiesModel';
 import { EntitySchema } from '../app/api/entities/entityType';
-import { ThesaurusSchema } from '../app/api/thesauris/dictionariesType';
 import { PropertySchema } from '../app/shared/commonTypes';
+import { ThesaurusSchema } from '../app/shared/thesaurusType';
 
 const { limit, recompute, overwrite, mode, model: fixedModel, sharedIds: sharedIdsStr } = yargs
   .option('limit', { default: 1000000 })
